test: добавить интеграционные тесты, исправить баги

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -19,6 +19,9 @@ public class UserService {
         if (userId <= 0 || friendId <= 0) {
             throw new EntityNotFoundException("Invalid User ID");
         }
+        if (userId == friendId) {
+            throw new IllegalArgumentException("Users cannot add themselves to friends");
+        }
         userStorage.addFriend(userId, friendId);
     }
 
@@ -26,12 +29,19 @@ public class UserService {
         if (userId <= 0 || friendId <= 0) {
             throw new EntityNotFoundException("Invalid User ID");
         }
+        if (userId == friendId) {
+            throw new IllegalArgumentException("Users cannot add themselves to friends");
+        }
         userStorage.deleteFriend(userId, friendId);
     }
 
     public List<User> getCommonFriends(long user1Id, long user2Id) {
         if (user1Id <= 0 || user2Id <= 0) {
             throw new EntityNotFoundException("Invalid User ID");
+        }
+        if (user1Id == user2Id) {
+            throw new IllegalArgumentException("Cannot pass a pair of same ids" +
+                    " (Suggest calling getFriends(long id) method instead)");
         }
         return userStorage.getCommonFriends(user1Id, user2Id);
     }
